Class of graphs where all nodes have even degree

This class is an extension of graphNEL-class. For graphs of this class, euler tours may always be constructed. Objects of this class should be created by mk_even_graph

Slots

This class has all slots from graphNEL-class plus:

dummy_node:

Object of class "character"

extra_edges:

Object of class "character"

weighted:

Object of class "logical"

% \item{\code{nodes}:}{Object of class \code{"vector"} ~~ } % \item{\code{edgeL}:}{Object of class \code{"list"} ~~ } % \item{\code{edgemode}:}{Object of class \code{"character"} ~~ } % \item{\code{edgeData}:}{Object of class \code{"attrData"} ~~ } % \item{\code{nodeData}:}{Object of class \code{"attrData"} ~~ } % \item{\code{renderInfo}:}{Object of class \code{"renderInfo"} ~~ } % \item{\code{graphData}:}{Object of class \code{"list"} ~~ }

Extends

Class graphNEL-class, directly. Class graph-class, by class "graphNEL", distance 2.

Methods

is_even_graph

signature(g = "graphNEL"): checks whether a graph has all nodes of even degree.

is_even_graph

signature(g = "even_graph"): always TRUE.

Author

C.B. Hurley and R.W. Oldford

References

see overview

Examples

showClass("even_graph")

#> Class "even_graph" [package "PairViz"]
#> 
#> Slots:
#>                                                                               
#> Name:   dummy_node extra_edges    weighted       nodes       edgeL    edgeData
#> Class:   character   character     logical      vector        list    attrData
#>                                           
#> Name:     nodeData  renderInfo   graphData
#> Class:    attrData  renderInfo        list
#> 
#> Extends: 
#> Class "graphNEL", directly
#> Class "graph", by class "graphNEL", distance 2
#> Class "graphBase", by class "graphNEL", distance 3