Function to overlay confidence intervals on the current plot.

Overlays confidence intervals on the current plot. Also draws a right hand axis, a horizontal broken line at zero, and marks the significant comparisons with an arrow, i.e. the CIs that do not intersect zero.

overlayCI(cis, xpos=NULL,ci.cols = NULL, ci.ex = 2, ci.ocol = "grey40", 
p.col = "grey40", pch = 1, sig.col = "red", sig.lwd = 1, yusr = NULL,
ci.label="Differences",ci.cex=0.5,arrow.length=0.1,...)

Arguments

cis
A matrix containing the confidence intervals. Each row corresponds to a different comparison, the first column is the estimated mean, and succesive pairs of columns give the lower and upper limits for different confidence levels.

ci.cols
A vector of colours, one colour for each confidence level. Defaults to shades of grey.

ci.ex
Controls confidence interval line width.

xpos
Horizonal positions where CIs are drawn. Defaults to 1.5,2.5,3.5,..

ci.ocol
Colour of zero line.

p.col
Colour of point used for CI centre.

pch
Symbol used for CI centre.

sig.col
Colour of arrow marking significant comparisons.

sig.lwd
Width of arrow marking significant comparisons.

yusr
Specifies the vertical par(usr) .Defaults to max and min.

ci.label
Label drawn on right margin.

ci.cex
Controls size of CI mean point symbol.

arrow.length
Controls size arrow at right hand axis.

...
Ignored

References

see overview

Author

C.B. Hurley and R.W. Oldford

Note

This function is called by mc_plot

See also